refactor: baseballGuessResponse 클래스 package 변경
