+++ src/main/java/com/keeper/homepage/domain/game/application/BaseballService.kt
@@ -1,7 +1,7 @@
 package com.keeper.homepage.domain.game.application
 
 import com.keeper.homepage.domain.game.dto.BaseballResult
-import com.keeper.homepage.domain.game.dto.req.BaseballGuessResponse
+import com.keeper.homepage.domain.game.dto.res.BaseballGuessResponse
 import com.keeper.homepage.domain.game.dto.res.GameInfoByMemberResponse
 import com.keeper.homepage.domain.member.entity.Member
 import com.keeper.homepage.global.error.BusinessException
